Changed collector into path specification extractor #338
