+++ tests/test_lib.py
@@ -1,8 +1,33 @@
 # -*- coding: utf-8 -*-
 """Shared functions and classes for testing."""
 
+import os
 import shutil
 import tempfile
+import unittest
+
+
+class BaseTestCase(unittest.TestCase):
+  """The base test case."""
+
+  _TEST_DATA_PATH = os.path.join(os.getcwd(), u'test_data')
+
+  # Show full diff results, part of TestCase so does not follow our naming
+  # conventions.
+  maxDiff = None
+
+  def _GetTestFilePath(self, path_segments):
+    """Retrieves the path of a test file relative to the test data directory.
+
+    Args:
+      path_segments: the path segments inside the test data directory.
+
+    Returns:
+      A path of the test file.
+    """
+    # Note that we need to pass the individual path segments to os.path.join
+    # and not a list.
+    return os.path.join(self._TEST_DATA_PATH, *path_segments)
 
 
 class TempDirectory(object):
